Move player authentication to StartMenuViewModel
